finalizing changes for watchOS
